+++ src/git/append.ts
@@ -1,5 +1,5 @@
 import { GitCommandsInstance } from "./instance";
-
+import { CacheManager } from "../manager/exports";
 export interface AppendNoteParameters {
   repositoryPath: string;
   commitHash: string;
@@ -7,11 +7,15 @@ export interface AppendNoteParameters {
 }
 
 export class AppendNote extends GitCommandsInstance {
+  private cache: CacheManager;
+
   constructor() {
     super(); // Call the constructor of the parent class
+    this.cache = CacheManager.getInstance();
   }
 
   public async command(parameter: AppendNoteParameters): Promise<void> {
+    // TODO: remove this check, only need to set repository with parameter interface
     if (!this.repositoryPath) {
       this.setRepositoryPath(parameter.repositoryPath);
     }
@@ -24,7 +28,8 @@ export class AppendNote extends GitCommandsInstance {
       "append",
       "-m",
       parameter.message,
-      parameter.commitHash];
+      parameter.commitHash,
+    ];
     await this.git
       .raw(cmdList)
       .then(async () => {
@@ -33,6 +38,10 @@ export class AppendNote extends GitCommandsInstance {
           parameter.commitHash,
           fullNote,
           parameter.repositoryPath
+        );
+        await this.cache.loadNoteDetails(
+          parameter.repositoryPath,
+          parameter.commitHash
         );
       })
       .catch((error) => {