+++ CHANGELOG.md
@@ -9,8 +9,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 
 - [#42](https://github.com/jrosco/vscode-git-notes/pull/42) ❇️ Added a new edit window class for git notes editor
-- [#43](https://github.com/jrosco/vscode-git-notes/pull/43) ❇️ Add `append`` command and refactor some code to support this new command
+- [#43](https://github.com/jrosco/vscode-git-notes/pull/43) ❇️ Add `append` command and refactor some code to support this new command
 - [#45](https://github.com/jrosco/vscode-git-notes/pull/45) 🔊 Add `deprecated` log service for deprecated classes and methods
+
+### Changed
+
+- [#44](https://github.com/jrosco/vscode-git-notes/pull/44) ❇️ Moved old git `interface` class to new git manager classes
+
 
 ## [0.2.1] - 2023-08-07